Forgot to remove the "Copy Of Player.cs" from the sln before uplaoding, sorry :D
Started work on the Group Database, going well so far :D
Started the Gooey, but there is nothing in it yet.

+++ MCFrog/Player/Player.cs
@@ -159,7 +159,7 @@ namespace MCFrog
 			PreLoader.PDB pdb = PreLoader.PDB.Find(Username.Trim().ToLower());
 			if(pdb == null)
 			{
-				var dbData = new object[] {Username, "", _ip, (byte) 0, (byte) 0, false, false};
+				var dbData = new object[] {Username, "", _ip, (byte) 0, 0, false, false};
 				UID = Server.users.NewRow(dbData);
 				new PreLoader.PDB(UID, dbData);
 				Server.Log(UID + " is this players UID :D", LogTypesEnum.Debug);
